Move FieldPropagatorTestBase in a separate header and reuse for user defined field tests
